requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
boto3==1.34.117
botocore==1.34.117
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
multidict==6.0.5
openai==1.31.1
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
